backend/requirements.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Data processing
numpy==1.24.3
pandas==2.0.3
ta==0.10.2

# WebSocket
websockets==12.0

# MT5 Linux compatibility - CORRECT VERSION
mt5linux==1.0.3

# Optional performance
orjson==3.9.7
httptools==0.6.0
